+++ connect-for-commerce-server-only/src/main/java/to/be/renamed/module/EcomConnectProjectApp.java
@@ -2,12 +2,14 @@ package to.be.renamed.module;
 
 import to.be.renamed.module.projectconfig.access.ProjectAppConfigurationService;
 import to.be.renamed.module.projectconfig.model.BridgeConfig;
+import to.be.renamed.module.projectconfig.model.CacheConfig;
 import to.be.renamed.module.projectconfig.model.GeneralConfig;
 import to.be.renamed.module.projectconfig.model.ProjectAppConfiguration;
 import to.be.renamed.module.projectconfig.model.ReportConfig;
 import to.be.renamed.module.setup.CaasIndexCreator;
 import com.espirit.moddev.components.annotations.ProjectAppComponent;
-
+import com.google.gson.JsonObject;
+import com.google.gson.JsonParser;
 import de.espirit.common.base.Logging;
 import de.espirit.common.tools.Strings;
 import de.espirit.firstspirit.io.FileHandle;
@@ -21,13 +23,10 @@ import java.io.InputStreamReader;
 import java.util.Arrays;
 import java.util.Properties;
 
-import com.google.gson.JsonObject;
-import com.google.gson.JsonParser;
-
 @ProjectAppComponent(
-    name = ProjectAppHelper.PROJECT_APP_NAME,
-    displayName = ProjectAppHelper.PROJECT_APP_NAME,
-    configurable = EcomConnectProjectConfig.class
+        name = ProjectAppHelper.PROJECT_APP_NAME,
+        displayName = ProjectAppHelper.PROJECT_APP_NAME,
+        configurable = EcomConnectProjectConfig.class
 )
 public class EcomConnectProjectApp implements ProjectApp {
 
@@ -59,18 +58,25 @@ public class EcomConnectProjectApp implements ProjectApp {
             final Properties props = loadOldConfiguration();
 
             final GeneralConfig generalConfig = new GeneralConfig(props.getProperty("bridge.content-creator-extension"),
-                                                                  Boolean.parseBoolean(props.getProperty("general.use-content-creator-extension")),
-                                                                  Boolean.parseBoolean(props.getProperty("checkbox.disable-bridge-page-creation")));
-            final BridgeConfig
-                bridgeConfig =
-                new BridgeConfig(props.getProperty("bridge.api-url"), props.getProperty("bridge.username"), props.getProperty("bridge.password"));
+                    Boolean.parseBoolean(props.getProperty("general.use-content-creator-extension")),
+                    Boolean.parseBoolean(props.getProperty("checkbox.disable-bridge-page-creation")));
+
+            Logging.logError(props.toString(), getClass());
+
+            final BridgeConfig bridgeConfig = new BridgeConfig(
+                    props.getProperty("bridge.api-url"),
+                    props.getProperty("bridge.username"),
+                    props.getProperty("bridge.password"),
+                    CacheConfig.fromStrings(
+                            props.getProperty("cache.size"),
+                            props.getProperty("cache.age")));
 
             int categoryLevels = 3;
             try {
                 categoryLevels = Integer.parseInt(props.getProperty("report.category-levels"));
             } catch (final NumberFormatException nfe) {
                 Logging.logInfo("No value for property report.category-levels found in old configuration file. Migrating to default value: 3",
-                                getClass());
+                        getClass());
             }
 
             int productLevels = 3;
@@ -78,7 +84,7 @@ public class EcomConnectProjectApp implements ProjectApp {
                 productLevels = Integer.parseInt(props.getProperty("report.product-levels"));
             } catch (final NumberFormatException nfe) {
                 Logging.logInfo("No value for property report.product-levels found in old configuration file. Migrating to default value: 3",
-                                getClass());
+                        getClass());
             }
 
             final ReportConfig reportConfig = new ReportConfig(categoryLevels, productLevels);
@@ -99,16 +105,16 @@ public class EcomConnectProjectApp implements ProjectApp {
 
                 final GeneralConfig generalConfig = projectAppConfiguration.getGeneralConfig();
                 final BridgeConfig
-                    bridgeConfig =
-                    new BridgeConfig(migratedBridgeApiUrl, projectAppConfiguration.getBridgeConfig().getBridgeUsername(),
-                                     projectAppConfiguration.getBridgeConfig().getBridgePassword());
+                        bridgeConfig =
+                        new BridgeConfig(migratedBridgeApiUrl, projectAppConfiguration.getBridgeConfig().getBridgeUsername(),
+                                projectAppConfiguration.getBridgeConfig().getBridgePassword(), projectAppConfiguration.getBridgeConfig().getCacheConfig());
                 final ReportConfig reportConfig = projectAppConfiguration.getReportConfig();
 
                 final ProjectAppConfiguration migratedConfiguration = new ProjectAppConfiguration(generalConfig, bridgeConfig, reportConfig);
 
                 storeMigratedConfiguration(migratedConfiguration);
                 Logging.logInfo("Project app configuration for project with ID " + environment.getProjectId() + " successfully migrated!",
-                                getClass());
+                        getClass());
             } else {
                 Logging.logInfo("No value for property bridge.api-url found in old configuration file. Leaving value empty.", getClass());
             }
@@ -118,8 +124,8 @@ public class EcomConnectProjectApp implements ProjectApp {
 
     protected void storeMigratedConfiguration(final ProjectAppConfiguration projectAppConfiguration) {
         final ProjectAppConfigurationService
-            projectAppConfigurationService =
-            ServiceFactory.getProjectAppConfigurationService(environment.getBroker());
+                projectAppConfigurationService =
+                ServiceFactory.getProjectAppConfigurationService(environment.getBroker());
         projectAppConfigurationService.storeConfiguration(projectAppConfiguration);
     }
 
@@ -138,16 +144,16 @@ public class EcomConnectProjectApp implements ProjectApp {
     protected ProjectAppConfiguration loadOldConfigurationFromJson() {
         final FileSystem<? extends FileHandle> fs = environment.getConfDir();
         final ProjectAppConfigurationService
-            projectAppConfigurationService =
-            ServiceFactory.getProjectAppConfigurationService(environment.getBroker());
+                projectAppConfigurationService =
+                ServiceFactory.getProjectAppConfigurationService(environment.getBroker());
         ProjectAppConfiguration projectAppConfiguration = projectAppConfigurationService.loadConfiguration();
         return projectAppConfiguration;
     }
 
     protected JsonObject loadOldJsonConfiguration() {
         final ProjectAppConfigurationService
-            projectAppConfigurationService =
-            ServiceFactory.getProjectAppConfigurationService(environment.getBroker());
+                projectAppConfigurationService =
+                ServiceFactory.getProjectAppConfigurationService(environment.getBroker());
         ProjectAppConfiguration projectAppConfiguration = projectAppConfigurationService.loadConfiguration();
         projectAppConfigurationService.storeConfiguration(projectAppConfiguration);
         final FileSystem<? extends FileHandle> fs = environment.getConfDir();